requirements.txt
browsercookie
bs4
ebooklib
requests
selenium
pyyaml